+++ include/semaphore.h
@@ -13,6 +13,6 @@ void sem_wait_(sem_t_ *s);
 
 void sem_post_(sem_t_ *s);
 
-int sem_get_value_(sem_t_ *s);
+long long int sem_get_value_(sem_t_ *s);
 
 #endif
